Fixes #93 (#94)
Addresses the integer overflow that could occur for large values of _ttl by first extending it to long before multiplying.

+++ src/main/java/javax/jmdns/impl/DNSRecord.java
@@ -126,7 +126,7 @@ public abstract class DNSRecord extends DNSEntry {
      */
     long getExpirationTime(int percent) {
         // ttl is in seconds the constant 10 is 1000 ms / 100 %
-        return _created + (percent * _ttl * 10L);
+        return _created + (percent * ((long)_ttl) * 10L);
     }
 
     /**
